wallet: skip threshold check during fee estimation
We have to skip the threshold check during fee estimation because we
might not have the necessary signers until the wallet actually signs the
transactions.

By design, the Torus login key only exists in memory in Sequence wallet
and cannot generally be assumed to be available. However the Torus login
key might be required in order to transact on other non-auth chains,
because the wallet config might not recognize the current session's
signing key. In these cases, the Torus key is retrieved when the user
confirms the transaction, which happens after fee estimation. So the
wallet might not meet the threshold during fee estimation despite
meeting it at confirmation time.